fix: 终端本地连接方式修改
