+++ frontend/src/routers/modules/host.ts
@@ -29,27 +29,7 @@ const hostRouter = {
         {
             path: '/hosts/terminal',
             name: 'Terminal',
-            component: () => import('@/views/host/terminal/terminal/index.vue'),
-            meta: {
-                title: 'menu.terminal',
-                keepAlive: true,
-            },
-        },
-        {
-            path: '/hosts/host',
-            name: 'TerminalHost',
-            hidden: true,
-            component: () => import('@/views/host/terminal/host/index.vue'),
-            meta: {
-                title: 'menu.terminal',
-                keepAlive: true,
-            },
-        },
-        {
-            path: '/hosts/command',
-            name: 'TerminalCommand',
-            hidden: true,
-            component: () => import('@/views/host/terminal/command/index.vue'),
+            component: () => import('@/views/host/terminal/index.vue'),
             meta: {
                 title: 'menu.terminal',
                 keepAlive: true,
